+++ simple_migration_test.py
@@ -26,7 +26,7 @@ def test_bayesian_tpe_import():
 def test_hpo_utils_import():
     """Test that we can import the migrated HPO utils."""
     try:
-        from src.utils.ml_common.optimization.hpo_utils import HyperparameterOptimization
+        from src.utils.nas_tas.advanced_hpo_utils import HyperparameterOptimization
         print("✅ HPO Utils import successful")
         return True
     except Exception as e:
