+++ src/training/steps/market_analysis/nas_modeling/core/neural_odes.py
@@ -87,7 +87,7 @@ except ImportError as e:
 # ML Common utilities
 try:
     from src.utils.ml_common.validation.unified_cv import perform_cross_validation
-    from src.utils.ml_common.optimization.hpo_utils import HyperparameterOptimizer
+    from src.utils.nas_tas.advanced_hpo_utils import HyperparameterOptimization as HyperparameterOptimizer
     from src.utils.ml_common.common_operations import get_ml_common_operations
     ML_COMMON_AVAILABLE = True
 except ImportError:
